+++ src/routes-test/Calendar.test.tsx
@@ -4,7 +4,7 @@
 import '@testing-library/jest-dom';
 import { render, screen, fireEvent, waitFor } from '@testing-library/react';
 import { BrowserRouter } from 'react-router-dom';
-import { Calendar } from './Calendar';
+import { Calendar } from '../routes/Calendar';
 import { UserContext } from '../util/userContext';
 import { getDocs, Timestamp } from 'firebase/firestore';
