+++ src/routes-test/ForgetMeNot.test.tsx
@@ -3,7 +3,7 @@
  */
 import '@testing-library/jest-dom';
 import { render, screen, fireEvent, waitFor } from '@testing-library/react';
-import ForgetMeNot from './ForgetMeNot';
+import ForgetMeNot from '../routes/ForgetMeNot';
 import { BrowserRouter } from 'react-router-dom';
 import { sendPasswordResetEmail } from 'firebase/auth';
 import { auth } from '../firebase/firebaseConfig';
