+++ src/routes-test/LoginIntegration.test.tsx
@@ -4,9 +4,9 @@
 import '@testing-library/jest-dom';
 import { render, screen, fireEvent, waitFor } from '@testing-library/react';
 import { MemoryRouter, Routes, Route } from 'react-router-dom';
-import Login from './Login';
-import ForgetMeNot from './ForgetMeNot';
-import { YourProfile } from './YourProfile';
+import Login from '../routes/Login';
+import ForgetMeNot from '../routes/ForgetMeNot';
+import { YourProfile } from '../routes/YourProfile';
 import { UserContext } from '../util/userContext';
 import { ToastContainer } from 'react-toastify';
 import { FirebaseError } from 'firebase/app';
